Updating routing tests to support new Route Message
The scripts which originaly was from the Routing were moved to RoutingV4 and Routing
was updated with new scripts using the Route message. The new tests for the version 4.3
is being skiped for any language but java since it's the only implementing the new message.

The support for the new message was added in scripting.Bolt4x3Script and register to be
used only in the version 4.3