Skip new retry tests on Python
Python sends a ROLLBACK on failure and Python also handles 'not a
leader' and 'forbidden read only database' different when bolt scheme is
in use. This is kind of correct, these errors should not happen on bolt
scheme but requires the tests to be rewritten to take routing into
account and that part differs A LOT between different drivers. On 4.3
the differences are a lot smaller...