Routing: Fix try_count verification
The drivers differs a bit about when it should or not call the transaction function.
The java driver only call the transaction function if the connection is fully setup, meanwhile the JS driver delay the connection resoltion until the last moment.

Thus the try_count is not a good way to undirect check if the driver try to call all the servers needed. Ideally, It should have a way to check if the full script was called,
but since it's not possible at this point, the adopt altertanative was checking for the occorrance of some events which could indicates the script was run on the stub server execution log.