Enabled 2 tests for java driver where router is also a reader
The java driver uses separate connections for routers and readers/writers when readers/writers are addressed by domain names in routing table. However, it will aim to use the same connection if IP addresses are used and there is no ambiguity involved. Since the purpose of these tests in not to test DNS resolution, it is suggested to switch them to IP-based addressing.