Skipping by capability/feature
This change provide a way to skip one or more tests using features names to make it easier to understand why the test is being skipped, specially in the case it is testing new features by avoid the if `list of language` approach. It does not replace the skipped lists implemented for each driver testkit-backend since the skip reason could be different.

The usage of the RunTest message to fetch capabilities was done to keep a simple and backwards compatible way to use the fetch this information.