Avoid using !:AUTO GOODBYE (#193)
This makes sure that server.done() actually asserts that the script has been
fully played. With auto GOODBYE the driver can reach the end of the script at
any point by sending a GOODBYE which is likely to happen on driver.close() which
should be called before checking server.done() to avoid race conditions in which
the server has not yet transmitted all received messages to TestKit.