Adding explicit auto lines to stubscript (#194)
`!: AUTO <SOME_COMMAND>` has the big drawback that it allows driver to send
the specified message at any given time. This can lead to tests passing that
actually shouldn't. The biggest pitfall is `!: AUTO GOODBYE` which more or less
renders `server.done()` useless as it allows the driver at any point in the
script to jump to the end by sending `GOODBYE` and making the server shut down
this way.
Therefore, I'm introducing a new line type: auto lines. They look like this

    A: RESET

This line will match the `RESET` message regardless of the fields attached to
the message (just like `!: AUTO` does and send the auto response (the same that
`!: AUTO` sends for each message).

Moreover, I introduce shorthand notation:

    ?: RESET

    *: RESET

    +: RESET

which will internally be expanded to

    {?
        A: RESET
    ?}

    {*
        A: RESET
    *}

    {+
        A: RESET
    +}

respectively.