Simulate SSR-enabled cluster with NotALeader failure
When a cluster enables SSR, it sends a routing table with a single
entry. When a "cluster-retriable error" such as "not a leader"
error occurs, it may happen that the middleware sitting between
the driver and the cluster keeps the failing connection alive.
Because the routing table always points to the same address, the
driver may reuse the same connection that previously failed and
therefore retry and fail again.

This test makes sure that drivers establish a new connection upon
such errors.

Signed-off-by: Florent Biville <florent.biville@neo4j.com>