Capture backend log to the artifacts
The backend logs were not captured by testkit to the artificats. This task was being done by the `backend.py` in the driver repository.

The problem with this approach is that sometimes the backend can not have access to the work directory and not be able log it. Also, if there is a bug in the `backend.py` the error logging about what's happening there will not be printed and recorded in the artifacts.