Introduce Support for Websocket in Bolt Stub (#351)
The support for Websockets in the Bolt Stub is need for testing the javascript driver in the browser.

The socket negotiation is done during the `BoltStubRequestHandler.setup` creating the wire using the `create_wire` function. This function handles the WebSocket negotiation and return the Wire with the correct socket negotiated.

The `WebSocket` implements read and receive data using the websockets frames and handle the control frames. This class doesn't not fulfil the whole `rfc6455` implementation because it always treat the payload as binary (which is the only kind of payload we use). The `negotiate_protocol` function also doesn't implement the whole `rfc6455` for negotiate websocket, It doesn't support extensions since It isn't needed for BoltStub use case.

PS: `RegularSocket` and `WebSocket` only implement and test the methods used by `Wire`