Add tests to check Result's scope
Result should raise an exception when the user tries to either access records
after consuming the result or after the transaction that owns the result has
been closed.

Ideally those two exceptions are not the same or have different error messages
to help the user in identifying the root cause and in fixing it.