Add tests for Bolt fix for broken DateTime encoding (#470)
The structures with signature `0x46` and `0x66` are being replaced by `0x49` and `0x69`. This new structures changes the meaning of seconds and nano seconds from `adjusted Unix epoch`  to `UTC`.

This changes have with goal of avoiding un-existing or ambiguous ZonedDateTime to be received or sent over Bolt.

Bolt v4.3 and v4.4 were patched to support this feature if the server supports the patch.

Co-authored-by: Antonio Barcelos <antonio.barcelos@neo4j.com>
Co-authored-by: grant lodge <6323995+thelonelyvulpes@users.noreply.github.com>
Co-authored-by: Florent Biville <florent.biville@neo4j.com>
Co-authored-by: Dmitriy Tverdiakov <dmitriy.tverdiakov@neo4j.com>