Make more error codes fail fast during the rediscovery (#536)
(Most) drivers will try really hard to fetch a valid RT. This entails going to the next know router if the previous one failed. In this process many errors that are considered unrecoverable in the transaction context are considered retryable.
However, a hand full of errors should not be retried to improve user experience. Namely those that are caused by malformed user-input. Instead of asking all routers if they can make sense of it and ultimately failing with a generic routing update error, it's better to fail fast and give the raw error of the server to the user.

Currently, these errors are
  - `Neo.ClientError.Database.DatabaseNotFound`
 - `Neo.ClientError.Transaction.InvalidBookmark`
 - `Neo.ClientError.Transaction.InvalidBookmarkMixture`

Added three more:
 - `Neo.ClientError.Statement.ArgumentError`, which happens on failed attempts of impersonation (user or permissions missing)
 - `Neo.ClientError.Request.Invalid`, which happens when trying  things like impersonating an integer
 - `Neo.ClientError.Statement.TypeError`, which happens when try to use a number as database name in protocol versions which don't support the `ROUTE` message.

Co-authored-by: Rouven Bauer <rouven.bauer@neo4j.com>