Stub server: conditional blocks and python lines
Adding new features to stub script

1. Simple Block

Allows you to group together any set of lines

```
{{
    C: HELLO
    S: ...
}}
```

Not very useful on it's own but meant to be used with

2. Conditional Block

Allows you to evaluate any Python code and branch accordingly

```
IF: some_condition == 123
{{
    ....
}}
ELIF: some < other < condition
{{
    ...
}}
ELSE:
{{
    ...
}}

As you might expect, you can have any number of `ELIF`s (including 0) and the
`ELSE` is optional.

3. Python Lines

```
PY: foo = 1 + 2
```

Executes arbitrary Python code. This is thread safe (for usage with concurrent
connections), and the variable space is shared across all connections.

It does not support multi-line expressions!

4. Python Bang Line

A new bang line is introduced. It is very similar to 3. Python Lines, but only
gets executed once (as opposed to per connection/restart). This is for instance
useful to initialize variables.

```
!: PY some_python = "statement"
```