Introduce Bolt Protocol 5.1
Implements the changes need for supporting Bolt v5.1 without the support for user switching and re-auth.

The goal of this changes is verify the support for the new bolt  protocol version in the drivers and the `LOGON` message implementation.

The `LOGOFF` message is not exercised since there isn't any use case for it.

Co-Authored-By: Rouven Bauer <rouven.bauer@neo4j.com>