Fix issues with linter
The warning should be log with stacktrace equals to 2. See B028.

The except block in wait_for_port was triggering B030 since the construction was not quite right.
Changing it to `except (OSError, ConnectionRefusedError):` caused a redundance since `ConnectionRefusedError` extends `OSError`,
so the except block could be reduce to treat only the base class.

ConnectionRefusedError