Adjust tx lifetime tests for Python (#598)
Currently, the tests assert that the driver throws the same error with which
the other result failed. I personally don't like it because that error has
nothing to do with the result stream being used (but only the other failed
stream). So instead I decided to throw some generic `ResultFailedError` in
Python with the other error as context information for easier debugging.

However, all of this is fool proofing of the drivers. Users will never encounter
this error unless they're abusing the driver or doing really funky stuff with
it. So I don't have a strong opinion about what exact error we should throw.
The main focus should be on preventing the driver from sending invalid messages
to the server and ideally throwing any error that's not absolutely cryptic.